refactor: Code conflict handling
- com.akatsuki.pioms.frwarehouse.aggregate.FranchiseWarehouse 리포지토리로 변경
- 영속성 제공하는 JpaRepository 인터페이스 구현체
- 레포지토리 패턴의 일환으로 데이터 액세스 계층 조정
