+++ src/main/java/com/akatsuki/pioms/frwarehouse/repository/FranchiseWarehouseRepository.java
@@ -1,7 +1,7 @@
-package com.akatsuki.pioms.franchiseWarehouse.repository;
+package com.akatsuki.pioms.frwarehouse.repository;
 
 
-import com.akatsuki.pioms.franchiseWarehouse.aggregate.FranchiseWarehouse;
+import com.akatsuki.pioms.frwarehouse.aggregate.FranchiseWarehouse;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
