+++ src/main/java/com/akatsuki/pioms/order/vo/OrderVO.java
@@ -1,6 +1,6 @@
 package com.akatsuki.pioms.order.vo;
 
-import com.akatsuki.pioms.franchise.etc.DELIVERY_DATE;
+import com.akatsuki.pioms.franchise.aggregate.DELIVERY_DATE;
 import com.akatsuki.pioms.order.entity.OrderEntity;
 import com.akatsuki.pioms.order.etc.ORDER_CONDITION;
 import lombok.AllArgsConstructor;
@@ -9,7 +9,6 @@ import lombok.NoArgsConstructor;
 import lombok.ToString;
 
 import java.time.LocalDateTime;
-import java.util.Date;
 
 @Getter
 @ToString
